+++ angular/projects/wcs-angular/src/public-api.ts
@@ -9,5 +9,6 @@ export * from './lib/directives/checkbox-value-accessor.directive';
 export * from './lib/directives/switch-value-accessor.directive';
 export * from './lib/directives/input-value-accessor.directive';
 export * from './lib/directives/textarea-value-accessor.directive';
+export * from './lib/directives/counter-value-accessor.directive';
 
 export * from './lib/proxies';