+++ example/src/app/app.component.ts
@@ -28,6 +28,7 @@ const DEFAULT_TAB_KEY = 'formly';
         <wcs-tab itemKey="formly" header="Formly"></wcs-tab>
         <wcs-tab itemKey="formly-styling" header="Formly Styling"></wcs-tab>
         <wcs-tab itemKey="modal" header="Modal"></wcs-tab>
+        <wcs-tab itemKey="counter" header="Counter"></wcs-tab>
       </wcs-tabs>
       <ng-container [ngSwitch]="selectedTab">
         <app-input-example *ngSwitchCase="'input'"></app-input-example>
@@ -38,6 +39,7 @@ const DEFAULT_TAB_KEY = 'formly';
         <app-formly-example *ngSwitchCase="'formly'"></app-formly-example>
         <app-formly-styling-example *ngSwitchCase="'formly-styling'"></app-formly-styling-example>
         <app-modal-example *ngSwitchCase="'modal'"></app-modal-example>
+        <app-counter-example *ngSwitchCase="'counter'"></app-counter-example>
       </ng-container>
     </main>
   `,